Server/pyproject.toml
[tool.poetry]
name = "task"
version = "0.1.0"
description = ""
authors = ["phong"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.67.0"
SQLAlchemy = "^1.4.22"
mysqlclient = "^2.0.3"
PyJWT = "^2.1.0"

[tool.poetry.dev-dependencies]
uvicorn = "^0.14.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
